# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Банковская система управления картами
  description: |
    API для управления банковскими картами, переводами средств и пользователями.
    Система предоставляет безопасное управление картами с шифрованием данных
    и ролевым доступом (ADMIN/USER).
  version: 1.0.0
  contact:
    name: API Support
    email: support@bankcards.example.com
  license:
    name: Proprietary
    url: http://bankcards.example.com/license

servers:
  - url: http://localhost:8080
    description: Локальный сервер разработки
  - url: https://api.bankcards.example.com
    description: Продакшн сервер

tags:
  - name: Аутентификация
    description: Регистрация, авторизация и управление пользователями
  - name: Карты
    description: Управление банковскими картами
  - name: Переводы
    description: Переводы средств между картами
  - name: Баланс
    description: Операции с балансом карт

security:
  - bearerAuth: []

paths:
  /api/v1/auth/sign-up:
    post:
      tags:
        - Аутентификация
      summary: Регистрация пользователя
      description: Создание нового пользователя с ролью USER
      operationId: signUp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpRequest'
            example:
              username: "Jon"
              password: "password123321"
      responses:
        '200':
          description: Успешная регистрация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtAuthenticationResponse'
        '400':
          description: Невалидные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpErrorResponse'
        '409':
          description: Пользователь уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpErrorResponse'
      security: []

  /api/v1/auth/sign-in:
    post:
      tags:
        - Аутентификация
      summary: Авторизация пользователя
      description: Получение JWT токена для аутентификации
      operationId: signIn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignInRequest'
            example:
              username: "Jon"
              password: "password123321"
      responses:
        '200':
          description: Успешная авторизация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtAuthenticationResponse'
        '401':
          description: Неверные учетные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpErrorResponse'
      security: []

  /api/v1/auth/give-admin/{userId}:
    post:
      tags:
        - Аутентификация
      summary: Назначить роль ADMIN
      description: |
        Назначение роли ADMIN для существующего пользователя.
        Доступно только пользователям с ролью ADMIN.
      operationId: giveAdmin
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          description: ID пользователя для назначения роли ADMIN
      responses:
        '200':
          description: Роль успешно назначена
        '403':
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpErrorResponse'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpErrorResponse'

  /api/v1/card/create:
    post:
      tags:
        - Карты
      summary: Создание карты
      description: |
        Создание новой банковской карты.
        Доступно только пользователям с ролью ADMIN.
      operationId: createCard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCardRequest'
            example:
              cardNumber: "1234567890123456"
              clientId: 1
              expirationDate: "2026-12-31"
              balance: 1000.00
      responses:
        '201':
          description: Карта успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '400':
          description: Невалидные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpErrorResponse'
        '403':
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpErrorResponse'
        '409':
          description: Карта уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpErrorResponse'

  /api/v1/card/{cardId}:
    get:
      tags:
        - Карты
      summary: Получение карты по ID
      description: |
        Получение информации о конкретной карте.
        Пользователи могут получать только свои карты, администраторы - любые.
      operationId: getCard
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          description: ID карты
      responses:
        '200':
          description: Информация о карте
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '403':
          description: Нет доступа к карте
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpErrorResponse'
        '404':
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpErrorResponse'

  /api/v1/card/get-all:
    get:
      tags:
        - Карты
      summary: Получение всех карт
      description: |
        Получение всех карт с пагинацией.
        Доступно только пользователям с ролью ADMIN.
      operationId: getAllCards
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
          description: Номер страницы (начинается с 0)
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 20
          description: Количество элементов на странице
      responses:
        '200':
          description: Список карт
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/CardResponse'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
                  size:
                    type: integer
                  number:
                    type: integer
        '403':
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpErrorResponse'

  /api/v1/card/get-all/user/{userId}:
    get:
      tags:
        - Карты
      summary: Получение карт пользователя
      description: |
        Получение всех карт конкретного пользователя с пагинацией.
        Пользователи могут получать только свои карты, администраторы - любые.
      operationId: getAllCardsByUserId
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          description: ID пользователя
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
          description: Номер страницы (начинается с 0)
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 20
          description: Количество элементов на странице
      responses:
        '200':
          description: Список карт пользователя
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/CardResponse'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
                  size:
                    type: integer
                  number:
                    type: integer
        '403':
          description: Нет доступа к картам пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpErrorResponse'

  /api/v1/card/get-all/{cardNumber}:
    get:
      tags:
        - Карты
      summary: Получение карт по последним 4 цифрам
      description: |
        Поиск карт по последним 4 цифрам номера.
        Доступно только пользователям с ролью ADMIN.
      operationId: getAllCardsByCardNumber
      parameters:
        - name: cardNumber
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]{4}$'
          description: Последние 4 цифры номера карты
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
          description: Номер страницы (начинается с 0)
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 20
          description: Количество элементов на странице
      responses:
        '200':
          description: Список найденных карт
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/CardResponse'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
                  size:
                    type: integer
                  number:
                    type: integer
        '403':
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpErrorResponse'

  /api/v1/card/update/{cardId}:
    post:
      tags:
        - Карты
      summary: Обновление карты
      description: |
        Обновление статуса или баланса карты.
        Доступно только пользователям с ролью ADMIN.
      operationId: updateCard
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          description: ID карты
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCardRequest'
            example:
              status: "BLOCKED"
              balance: 1500.00
      responses:
        '200':
          description: Карта успешно обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '400':
          description: Невалидные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpErrorResponse'
        '403':
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpErrorResponse'
        '404':
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpErrorResponse'

  /api/v1/card/blocked/{cardId}:
    post:
      tags:
        - Карты
      summary: Запрос на блокировку карты
      description: |
        Пользователь может запросить блокировку своей активной карты.
        Карта переходит в статус REQUEST_FOR_BLOCKING.
      operationId: blockedCard
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          description: ID карты для блокировки
      responses:
        '200':
          description: Запрос на блокировку создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '400':
          description: Карта не активна или другая ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpErrorResponse'
        '403':
          description: Нет доступа к карте
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpErrorResponse'
        '404':
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpErrorResponse'

  /api/v1/card/balance/{cardId}:
    get:
      tags:
        - Баланс
      summary: Получение баланса карты
      description: |
        Получение текущего баланса конкретной карты.
        Пользователи могут получать баланс только своих карт.
      operationId: getCardBalance
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          description: ID карты
      responses:
        '200':
          description: Баланс карты
          content:
            application/json:
              schema:
                type: object
                properties:
                  balance:
                    type: number
                    format: decimal
                    example: 1500.50
        '403':
          description: Нет доступа к карте
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpErrorResponse'
        '404':
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpErrorResponse'

  /api/v1/card/transfer:
    post:
      tags:
        - Переводы
      summary: Перевод средств между картами
      description: |
        Перевод средств между двумя картами.
        Обе карты должны быть активны и принадлежать текущему пользователю.
        Для администраторов доступны любые переводы.
      operationId: transferBetweenCards
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardTransferRequest'
            example:
              fromCardId: 1
              toCardId: 2
              amount: 100.50
              description: "Перевод на покупки"
      responses:
        '200':
          description: Перевод успешно выполнен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardTransferResponse'
        '400':
          description: |
            - Недостаточно средств
            - Неверные ID карт
            - Карты не активны
            - Другая ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpErrorResponse'
        '403':
          description: Нет доступа к одной из карт
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpErrorResponse'
        '404':
          description: Одна из карт не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpErrorResponse'

components:
  schemas:
    # Запросы
    SignUpRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 2
          maxLength: 100
          description: Имя пользователя
          example: "Jon"
        password:
          type: string
          minLength: 8
          maxLength: 255
          description: Пароль
          example: "password123321"
      description: Запрос на регистрацию пользователя

    SignInRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 2
          maxLength: 100
          description: Имя пользователя
          example: "Jon"
        password:
          type: string
          minLength: 8
          maxLength: 255
          description: Пароль
          example: "password123321"
      description: Запрос на аутентификацию

    CreateCardRequest:
      type: object
      required:
        - cardNumber
        - clientId
        - expirationDate
        - balance
      properties:
        cardNumber:
          type: string
          pattern: '^[0-9]{16}$'
          description: Номер карты (16 цифр)
          example: "1234567890123456"
        clientId:
          type: integer
          format: int64
          minimum: 1
          description: ID владельца карты
          example: 1
        expirationDate:
          type: string
          format: date
          description: Дата окончания действия
          example: "2026-12-31"
        balance:
          type: number
          format: decimal
          minimum: 0
          description: Начальный баланс
          example: 1000.00
      description: Запрос на создание карты

    UpdateCardRequest:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/CardStatus'
          description: Новый статус карты
          example: "BLOCKED"
        balance:
          type: number
          format: decimal
          minimum: 0
          description: Новый баланс карты
          example: 1500.00
      description: Запрос на обновление карты

    CardTransferRequest:
      type: object
      required:
        - fromCardId
        - toCardId
        - amount
      properties:
        fromCardId:
          type: integer
          format: int64
          minimum: 1
          description: ID карты отправителя
          example: 1
        toCardId:
          type: integer
          format: int64
          minimum: 1
          description: ID карты получателя
          example: 2
        amount:
          type: number
          format: decimal
          minimum: 0.01
          description: Сумма перевода
          example: 100.50
        description:
          type: string
          description: Описание перевода
          example: "Перевод на покупки"
      description: Запрос на перевод средств

    # Ответы
    JwtAuthenticationResponse:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          description: JWT токен для аутентификации
          example: "eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJhZG1pbiIsImV4cCI6MTYyMjUwNj..."
      description: Ответ с токеном доступа

    CardResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор карты
          example: 1
        encryptedCardNumber:
          type: string
          description: Зашифрованный номер карты
          example: "h7/ywvSFSp9VbOzqOL4y4wUBh6DN5amHLLqwkatz5VM="
        clientId:
          type: integer
          format: int64
          description: ID владельца карты
          example: 1
        expirationDate:
          type: string
          format: date
          description: Дата окончания действия
          example: "2026-12-31"
        status:
          $ref: '#/components/schemas/CardStatus'
          description: Текущий статус карты
          example: "ACTIVE"
        balance:
          type: number
          format: decimal
          description: Текущий баланс
          example: 1500.50
      description: Ответ с информацией о карте

    CardTransferResponse:
      type: object
      properties:
        transactionId:
          type: integer
          format: int64
          description: ID транзакции
          example: 1
        fromCardId:
          type: integer
          format: int64
          description: ID карты отправителя
          example: 1
        toCardId:
          type: integer
          format: int64
          description: ID карты получателя
          example: 2
        amount:
          type: number
          format: decimal
          description: Сумма перевода
          example: 100.50
        fromCardNewBalance:
          type: number
          format: decimal
          description: Новый баланс карты отправителя
          example: 900.50
        toCardNewBalance:
          type: number
          format: decimal
          description: Новый баланс карты получателя
          example: 1100.50
        description:
          type: string
          description: Описание перевода
          example: "Перевод на покупки"
        timestamp:
          type: string
          format: date-time
          description: Время создания транзакции
          example: "2024-01-20T10:30:45.123Z"
      description: Ответ с информацией о переводе

    HttpErrorResponse:
      type: object
      properties:
        code:
          type: integer
          description: HTTP статус код
          example: 403
        type:
          type: string
          description: Тип ошибки
          example: "Forbidden"
        message:
          type: string
          description: Сообщение об ошибке
          example: "У вас нет прав для взаимодействия с картой другого пользователя"
      description: Ответ с информацией об ошибке

    # Enums
    CardStatus:
      type: string
      enum:
        - ACTIVE
        - REQUEST_FOR_BLOCKING
        - BLOCKED
        - EXPIRED
      description: Статус карты
      example: "ACTIVE"

    Role:
      type: string
      enum:
        - ROLE_USER
        - ROLE_ADMIN
      description: Роль пользователя
      example: "ROLE_USER"

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT токен, полученный при авторизации.
        Добавляйте токен в заголовок Authorization: "Bearer {token}"